Extract Witness types from babbage `Tx` module into new module `TxWits`
