+++ eras/babbage/impl/cardano-ledger-babbage.cabal
@@ -39,6 +39,7 @@ library
     Cardano.Ledger.Babbage.Tx
     Cardano.Ledger.Babbage.TxBody
     Cardano.Ledger.Babbage.TxInfo
+    Cardano.Ledger.Babbage.TxWits
     Cardano.Ledger.Babbage.Translation
     Cardano.Ledger.Babbage.Scripts
     Cardano.Ledger.Babbage.Collateral
